requirements.txt
asgiref==3.2.10
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
Django==3.0.8
enum-compat==0.0.3
future==0.18.2
idna==2.9
pycparser==2.20
python-telegram-bot==13.5
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.9
tronapi==3.1.8  # 新增 TRON API 库，用于钱包操作和交易
